De-dupe IdentityGetResponse classes for using in AssetReportGetResponse
